\name{MLL}
\docType{data}
\alias{MLL.A}
\alias{MLL.B}
\title{AffyBatch instances MLL.A and MLL.B}
\description{
These \code{\link[affy:AffyBatch-class]{AffyBatch}} objects contain a subset
of arrays from a large acute lymphoblastic leukemia (ALL) study.
}
\usage{
data(MLL.A)
data(MLL.B)
}
\format{Each are \code{\link[affy:AffyBatch-class]{AffyBatch}}
  containing 20 arrays.} 
\source{
  This package provides probe-level data for 20 HGU133A and 20 HGU133B
  arrays which are a subset of arrays from a large ALL study. The data
  is for the MLL arrays. This data was published in:

  Mary E. Ross, Xiaodong Zhou, Guangchun Song, Sheila A. Shurtleff,
  Kevin Girtman, W. Kent Williams, Hsi-Che Liu, Rami Mahfouz, Susana
  C. Raimondi, Noel Lenny, Anami Patel, and James R. Downing (2003)
  \emph{Classification of pediatric acute lymphoblastic leukemia by gene
  expression profiling} Blood 102: 2951-2959 
  }


\keyword{datasets}
